refactor(js): belgiumPhone validation to be consistent with the test
